+++ app/controllers/conversations.js
@@ -1,6 +1,8 @@
 const Conversation = require('../models/conversation.js')
 const validator = require('node-validator')
 const check = require('./payload-validator/conversations.js')
+const JWT = require('../jwt.js')
+const jwt = new JWT()
 
 /**
  * Conversations
@@ -24,16 +26,23 @@ class Conversations {
   create () {
     this.app.post('/conversations/create', validator.express(check), (req, res) => {
       try {
-        const oModel = new this.OModel(req.body)
-        
-        oModel.save().then(event => {
-          res.status(200).json(event || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          const oModel = new this.OModel(req.body)
+          
+          oModel.save().then(event => {
+            res.status(200).json(event || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -49,15 +58,22 @@ class Conversations {
   delete () {
     this.app.delete('/conversations/delete/:id', (req, res) => {
       try {
-        this.OModel.findOneAndDelete({_id: req.params.id})
-          .then(event => {
-            res.status(200).json(event || {})
-          }).catch(err => {
-            res.status(500).json({
-              'code': 500,
-              'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.OModel.findOneAndDelete({_id: req.params.id})
+            .then(event => {
+              res.status(200).json(event || {})
+            }).catch(err => {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
             })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
           })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -73,14 +89,21 @@ class Conversations {
   show () {
     this.app.get('/conversations/show/:id', (req, res) => {
       try {
-        this.OModel.findOne({_id: req.params.id}).then(event => {
-          res.status(200).json(event || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.OModel.findOne({_id: req.params.id}).then(event => {
+            res.status(200).json(event || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -96,16 +119,23 @@ class Conversations {
   update () {
     this.app.put('/conversations/update/:id', validator.express(check), (req, res) => {
       try {
-        this.OModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(doc || {})
-          }
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.OModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(doc || {})
+            }
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -121,16 +151,23 @@ class Conversations {
   list () {
     this.app.get('/conversations/list', (req, res) => {
       try {
-        this.OModel.find({}, function (err, result) {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(result)
-          }   
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.OModel.find({}, function (err, result) {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(result)
+            }   
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
