+++ app/controllers/messages.js
@@ -1,6 +1,8 @@
 const Message = require('../models/message.js')
 const validator = require('node-validator')
 const check = require('./payload-validator/messages.js')
+const JWT = require('../jwt.js')
+const jwt = new JWT()
 
 /**
  * Messages
@@ -24,16 +26,23 @@ class Messages {
   create () {
     this.app.post('/messages/create', validator.express(check), (req, res) => {
       try {
-        const msgModel = new this.MsgModel(req.body)
-        
-        msgModel.save().then(result => {
-          res.status(200).json(result || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          const msgModel = new this.MsgModel(req.body)
+          
+          msgModel.save().then(result => {
+            res.status(200).json(result || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -49,15 +58,22 @@ class Messages {
   delete () {
     this.app.delete('/messages/delete/:id', (req, res) => {
       try {
-        this.MsgModel.findOneAndDelete({_id: req.params.id})
-          .then(result => {
-            res.status(200).json(result || {})
-          }).catch(err => {
-            res.status(500).json({
-              'code': 500,
-              'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.MsgModel.findOneAndDelete({_id: req.params.id})
+            .then(result => {
+              res.status(200).json(result || {})
+            }).catch(err => {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
             })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
           })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -73,14 +89,21 @@ class Messages {
   show () {
     this.app.get('/messages/show/:id', (req, res) => {
       try {
-        this.MsgModel.findOne({_id: req.params.id}).then(result => {
-          res.status(200).json(result || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.MsgModel.findOne({_id: req.params.id}).then(result => {
+            res.status(200).json(result || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -96,16 +119,23 @@ class Messages {
   update () {
     this.app.put('/messages/update/:id', validator.express(check), (req, res) => {
       try {
-        this.MsgModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(doc || {})
-          }
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.MsgModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(doc || {})
+            }
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -121,16 +151,23 @@ class Messages {
   list () {
     this.app.get('/messages/list', (req, res) => {
       try {
-        this.MsgModel.find({}, function (err, result) {
-          if (err) {
-            return res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(result)
-          }   
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.MsgModel.find({}, function (err, result) {
+            if (err) {
+              return res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(result)
+            }   
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
