+++ app/controllers/surveys.js
@@ -1,6 +1,8 @@
 const Survey = require('../models/survey.js')
 const validator = require('node-validator')
 const check = require('./payload-validator/surveys.js')
+const JWT = require('../jwt.js')
+const jwt = new JWT()
 
 /**
  * Surveys
@@ -24,16 +26,23 @@ class Surveys {
   create () {
     this.app.post('/surveys/create', validator.express(check), (req, res) => {
       try {
-        const surveyModel = new this.SurveyModel(req.body)
-        
-        surveyModel.save().then(survey => {
-          res.status(200).json(survey || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          const surveyModel = new this.SurveyModel(req.body)
+          
+          surveyModel.save().then(survey => {
+            res.status(200).json(survey || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -49,15 +58,22 @@ class Surveys {
   delete () {
     this.app.delete('/surveys/delete/:id', (req, res) => {
       try {
-        this.SurveyModel.findOneAndDelete({_id: req.params.id})
-          .then(survey => {
-            res.status(200).json(survey || {})
-          }).catch(err => {
-            res.status(500).json({
-              'code': 500,
-              'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.SurveyModel.findOneAndDelete({_id: req.params.id})
+            .then(survey => {
+              res.status(200).json(survey || {})
+            }).catch(err => {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
             })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
           })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -73,14 +89,21 @@ class Surveys {
   show () {
     this.app.get('/surveys/show/:id', (req, res) => {
       try {
-        this.SurveyModel.findOne({_id: req.params.id}).then(survey => {
-          res.status(200).json(survey || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.SurveyModel.findOne({_id: req.params.id}).then(survey => {
+            res.status(200).json(survey || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -96,16 +119,23 @@ class Surveys {
   update () {
     this.app.put('/surveys/update/:id', validator.express(check), (req, res) => {
       try {
-        this.SurveyModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(doc || {})
-          }
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.SurveyModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(doc || {})
+            }
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -121,16 +151,23 @@ class Surveys {
   list () {
     this.app.get('/surveys/list', (req, res) => {
       try {
-        this.SurveyModel.find({}, function (err, result) {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(result)
-          }   
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.SurveyModel.find({}, function (err, result) {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(result)
+            }   
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
