+++ app/controllers/ticketings.js
@@ -1,7 +1,8 @@
 const Ticketing = require('../models/ticketing.js')
 const validator = require('node-validator')
 const check = require('./payload-validator/ticketings.js')
-
+const JWT = require('../jwt.js')
+const jwt = new JWT()
 /**
  * Ticketing 
  * @class
@@ -24,16 +25,23 @@ class Ticketings {
   create () {
     this.app.post('/ticketings/create', validator.express(check), (req, res) => {
       try {
-        const tickModel = new this.TickModel(req.body)
-        
-        tickModel.save().then(result => {
-          res.status(200).json(result || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          const tickModel = new this.TickModel(req.body)
+          
+          tickModel.save().then(result => {
+            res.status(200).json(result || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -49,15 +57,22 @@ class Ticketings {
   delete () {
     this.app.delete('/ticketings/delete/:id', (req, res) => {
       try {
-        this.TickModel.findOneAndDelete({_id: req.params.id})
-          .then(result => {
-            res.status(200).json(result || {})
-          }).catch(err => {
-            res.status(500).json({
-              'code': 500,
-              'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.TickModel.findOneAndDelete({_id: req.params.id})
+            .then(result => {
+              res.status(200).json(result || {})
+            }).catch(err => {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
             })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
           })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -73,14 +88,21 @@ class Ticketings {
   show () {
     this.app.get('/ticketings/show/:id', (req, res) => {
       try {
-        this.TickModel.findOne({_id: req.params.id}).then(result => {
-          res.status(200).json(result || {})
-        }).catch(err => {
-          res.status(500).json({
-            'code': 500,
-            'message': err
+        if (jwt.getToken(req.body.token)) {
+          this.TickModel.findOne({_id: req.params.id}).then(result => {
+            res.status(200).json(result || {})
+          }).catch(err => {
+            res.status(500).json({
+              'code': 500,
+              'message': err
+            })
           })
-        })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -96,16 +118,23 @@ class Ticketings {
   update () {
     this.app.put('/ticketings/update/:id', validator.express(check), (req, res) => {
       try {
-        this.TickModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
-          if (err) {
-            res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(doc || {})
-          }
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.TickModel.findOneAndUpdate({_id: req.params.id}, {$set: req.body}, {new: true}, (err, doc) => {
+            if (err) {
+              res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(doc || {})
+            }
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
@@ -121,16 +150,23 @@ class Ticketings {
   list () {
     this.app.get('/ticketings/list', (req, res) => {
       try {
-        this.TickModel.find({}, function (err, result) {
-          if (err) {
-            return res.status(500).json({
-              'code': 500,
-              'message': err
-            })
-          } else {
-            res.status(200).json(result)
-          }   
-        })
+        if (jwt.getToken(req.body.token)) {
+          this.TickModel.find({}, function (err, result) {
+            if (err) {
+              return res.status(500).json({
+                'code': 500,
+                'message': err
+              })
+            } else {
+              res.status(200).json(result)
+            }   
+          })
+        } else {
+          res.status(401).json({
+            'code': 401,
+            'message': 'Access Denied' 
+          })
+        }
       } catch (err) {
         res.status(500).json({
           'code': 500,
